Refactoring class structure and release enhancement
