+++ docker-compose.yml
@@ -33,9 +33,6 @@ services:
       - DENIED_EMAIL_ENABLED=${DENIED_EMAIL_ENABLED:-false}
       - COMMA_SEPARATED_EMAILS=${COMMA_SEPARATED_EMAILS:-dummyemail}
       - accessGrantEmailSubject=Testing
-      - IDP_PROVIDER_URI=https://staging.datastage.io
-      - FENCE_CLIENT_ID=${FENCE_CLIENT_ID:-no_fence_client_id}
-      - FENCE_CLIENT_SECRET=${FENCE_CLIENT_SECRET:-no_fence_client_secret}
       - MAIL_USERNAME=${MAIL_USERNAME:-emailuser@avillach.lab}
       - MAIL_PASSWORD=${MAIL_PASSWORD:-email_password}
       - TOS_ENABLED:false
@@ -47,8 +44,8 @@ services:
       - ./pic-sure-auth-services/src/main/resources/emailTemplates:${EMAIL_TEMPLATE_PATH:-/opt/jboss/wildfly/emailTemplates}
     networks:
       - public
-    #env_file:
-    #  - .env
+    env_file:
+      - .env
 
   picsureauthdb:
     build:
