--- fence-test-app/src/test/java/edu/harvard/hms/dbmi/avillach/picsure/auth/AppTest.java
@@ -1,20 +1,0 @@
-package edu.harvard.hms.dbmi.avillach.picsure.auth;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
