+++ pic-sure-auth-services/src/main/java/edu/harvard/hms/dbmi/avillach/auth/rest/AuthService.java
@@ -1,13 +1,17 @@
 package edu.harvard.hms.dbmi.avillach.auth.rest;
 
-import edu.harvard.dbmi.avillach.util.response.PICSUREResponse;
+import edu.harvard.hms.dbmi.avillach.auth.JAXRSConfiguration;
 import edu.harvard.hms.dbmi.avillach.auth.service.auth.AuthenticationService;
 import edu.harvard.hms.dbmi.avillach.auth.service.auth.AuthorizationService;
+import edu.harvard.hms.dbmi.avillach.auth.service.auth.FENCEAuthenticationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
-import javax.ws.rs.*;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
 import java.util.Map;
 
@@ -24,15 +28,20 @@ public class AuthService {
     @Inject
     AuthenticationService authenticationService;
 
+    @Inject
+    FENCEAuthenticationService fenceAuthenticationService;
+
     @POST
     @Path("/authentication")
     public Response authentication(Map<String, String> authRequest){
-        return authenticationService.getToken(authRequest);
+        logger.debug("authentication() starting...");
+        if (JAXRSConfiguration.idp_provider.equalsIgnoreCase("fence")) {
+            logger.debug("authentication() FENCE authentication");
+            return fenceAuthenticationService.getFENCEProfile(authRequest);
+        } else {
+            logger.debug("authentication() default authentication");
+            return authenticationService.getToken(authRequest);
+        }
     }
 
-    @POST
-    @Path("/fence-authentication")
-    public Response authenticationWithFence(Map<String, String> authRequest){
-        return authenticationService.getFENCEProfile(authRequest);
-    }
 }
